refactor: move db & storage adapter to lib folder
