+++ lib/storage-driver.ts
@@ -1,4 +1,4 @@
-import type { StorageDriver } from '@/utils/types'
+import type { StorageDriver } from '@/lib/types'
 import type { z } from 'zod'
 
 export function encodeCacheKey(key: string, version: string) {
@@ -13,7 +13,9 @@ export function defineStorageDriver<EnvSchema extends z.ZodTypeAny>(
   options: DefineStorageDriverOptions<EnvSchema>,
 ) {
   return () => {
+    // eslint-disable-next-line ts/no-unsafe-assignment
     const env = options.envSchema.parse(process.env)
+    // eslint-disable-next-line ts/no-unsafe-argument
     const driver = options.setup(env)
     return driver instanceof Promise ? driver : Promise.resolve(driver)
   }
